Update Expire | Enable User Profile
Some companies want their staff to be locked out for certain periods of time while a employee is on vacation, This will expire the account instead of disabling it. This will avoid any redflags on a system saying that account has been tampered with. Then it will re-enable the account and allow employee to resume regular duties right onschedule. The task will delete itself after use.